move generic config to common/txmgr/types
