+++ core/chains/evm/txmgr/reaper.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/sqlx"
 
+	txmgrtypes "github.com/smartcontractkit/chainlink/v2/common/txmgr/types"
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
 	"github.com/smartcontractkit/chainlink/v2/core/services/pg"
 	"github.com/smartcontractkit/chainlink/v2/core/utils"
@@ -17,7 +18,7 @@ import (
 // Reaper handles periodic database cleanup for Txm
 type Reaper struct {
 	db             *sqlx.DB
-	config         ReaperConfig
+	config         txmgrtypes.ReaperConfig
 	chainID        utils.Big
 	log            logger.Logger
 	latestBlockNum atomic.Int64
@@ -27,7 +28,7 @@ type Reaper struct {
 }
 
 // NewReaper instantiates a new reaper object
-func NewReaper(lggr logger.Logger, db *sqlx.DB, config ReaperConfig, chainID big.Int) *Reaper {
+func NewReaper(lggr logger.Logger, db *sqlx.DB, config txmgrtypes.ReaperConfig, chainID big.Int) *Reaper {
 	r := &Reaper{
 		db,
 		config,
