+++ core/chains/evm/txmgr/reaper_test.go
@@ -9,8 +9,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	txmgrtypes "github.com/smartcontractkit/chainlink/v2/common/txmgr/types"
+	txmgrmocks "github.com/smartcontractkit/chainlink/v2/common/txmgr/types/mocks"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr"
-	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr/mocks"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest"
 	configtest "github.com/smartcontractkit/chainlink/v2/core/internal/testutils/configtest/v2"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils/pgtest"
@@ -18,11 +19,11 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
 
-func newReaperWithChainID(t *testing.T, db *sqlx.DB, cfg txmgr.ReaperConfig, cid big.Int) *txmgr.Reaper {
+func newReaperWithChainID(t *testing.T, db *sqlx.DB, cfg txmgrtypes.ReaperConfig, cid big.Int) *txmgr.Reaper {
 	return txmgr.NewReaper(logger.TestLogger(t), db, cfg, cid)
 }
 
-func newReaper(t *testing.T, db *sqlx.DB, cfg txmgr.ReaperConfig) *txmgr.Reaper {
+func newReaper(t *testing.T, db *sqlx.DB, cfg txmgrtypes.ReaperConfig) *txmgr.Reaper {
 	return newReaperWithChainID(t, db, cfg, cltest.FixtureChainID)
 }
 
@@ -39,7 +40,7 @@ func TestReaper_ReapEthTxes(t *testing.T) {
 	oneDayAgo := time.Now().Add(-24 * time.Hour)
 
 	t.Run("with nothing in the database, doesn't error", func(t *testing.T) {
-		config := mocks.NewReaperConfig(t)
+		config := txmgrmocks.NewReaperConfig(t)
 		config.On("FinalityDepth").Return(uint32(10))
 		config.On("TxReaperThreshold").Return(1 * time.Hour)
 
@@ -53,7 +54,7 @@ func TestReaper_ReapEthTxes(t *testing.T) {
 	cltest.MustInsertConfirmedEthTxWithReceipt(t, borm, from, nonce, 5)
 
 	t.Run("skips if threshold=0", func(t *testing.T) {
-		config := mocks.NewReaperConfig(t)
+		config := txmgrmocks.NewReaperConfig(t)
 		config.On("TxReaperThreshold").Return(0 * time.Second)
 
 		r := newReaper(t, db, config)
@@ -65,7 +66,7 @@ func TestReaper_ReapEthTxes(t *testing.T) {
 	})
 
 	t.Run("doesn't touch ethtxes with different chain ID", func(t *testing.T) {
-		config := mocks.NewReaperConfig(t)
+		config := txmgrmocks.NewReaperConfig(t)
 		config.On("FinalityDepth").Return(uint32(10))
 		config.On("TxReaperThreshold").Return(1 * time.Hour)
 
@@ -78,7 +79,7 @@ func TestReaper_ReapEthTxes(t *testing.T) {
 	})
 
 	t.Run("deletes confirmed eth_txes that exceed the age threshold with at least EVM.FinalityDepth blocks above their receipt", func(t *testing.T) {
-		config := mocks.NewReaperConfig(t)
+		config := txmgrmocks.NewReaperConfig(t)
 		config.On("FinalityDepth").Return(uint32(10))
 		config.On("TxReaperThreshold").Return(1 * time.Hour)
 
@@ -105,7 +106,7 @@ func TestReaper_ReapEthTxes(t *testing.T) {
 	cltest.MustInsertFatalErrorEthTx(t, borm, from)
 
 	t.Run("deletes errored eth_txes that exceed the age threshold", func(t *testing.T) {
-		config := mocks.NewReaperConfig(t)
+		config := txmgrmocks.NewReaperConfig(t)
 		config.On("FinalityDepth").Return(uint32(10))
 		config.On("TxReaperThreshold").Return(1 * time.Hour)
